+++ src/app/stats/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { Suspense } from 'react';
-import { StatsMatrix } from '@/components/stats/StatsMatrix';
+import { StatsMatrix } from '@/components/stats';
 import { Loading } from '@/components/ui/Loading';
 
 export default function StatsPage() {
